+++ src/app/_dto/misskey-callback/user-info.dto.ts
@@ -1,4 +1,4 @@
-import type { MiUser } from '@/app/api/misskey-entities/user';
+import type { MiUser } from '@/app/api/_misskey-entities/user';
 
 export class misskeyUserInfoPayload {
   user: MiUser;
