+++ src/app/api/db/fetch-my-questions/route.ts
@@ -1,8 +1,8 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { verifyToken } from '../../functions/web/verify-jwt';
-import { sendApiError } from '@/utils/apiErrorResponse/sendApiError';
-import { GetPrismaClient } from '@/utils/getPrismaClient/get-prisma-client';
-import { RateLimiterService } from '@/utils/ratelimiter/rateLimiter';
+import { verifyToken } from '../../_utils/jwt/verify-jwt';
+import { sendApiError } from '@/app/api/_utils/apiErrorResponse/sendApiError';
+import { GetPrismaClient } from '@/app/api/_utils/getPrismaClient/get-prisma-client';
+import { RateLimiterService } from '@/app/api/_utils/ratelimiter/rateLimiter';
 
 export async function GET(req: NextRequest) {
   const prisma = GetPrismaClient.getClient();
