--- src/app/api/functions/web/errorResponse.ts
@@ -1,5 +1,0 @@
-import { NextResponse } from 'next/server';
-
-export function sendErrorResponse(code: number, message: string) {
-  return NextResponse.json({ status: code, message: message }, { status: code });
-}
